Deprecate export file
for more information, see https://pre-commit.ci

for more information, see https://pre-commit.ci
